requirements.txt
openai
python-chess
